Add user prompt loop and checkFileType function.
Within a while True loop, user is prompted for the file name to convert.
The file name is passed into the checkFileType function along with
the .csv extension as args. checkFileType is a reusable function that
accepts any file extension as an arg, returns True if valid and False if
invalid. If invalid, it also returns information to the user.

If checkFileType returns True, the program continues. If False, user
will be prompted again to enter the file name.

KeyboardInterrupt error handling and a sys.exit on successful
program completion also added.