Revise call of checkFileType, add checkFilePath.
If statement calling checkFileType changed to if not with a continue
to reprompt user for readability.

checkFilePath function added, which uses Path.home() and 'Downloads' to construct a
relative path to the file. It then checks to see if the path exists,
returns the path if valid, and displays a confirmation message to user.
If path is invalid, displays a message to user, does not return the path,
and if not statement triggers continue of reprompt.