Add getFieldnames function.
getFieldnames uses the file path and CSV DictReader to parse the file
and return a list of fieldnames from the header.

If the document is empty, a StopIteration error will be thrown. This is
handled by sending a message to the user in the terminal and triggering
a reprompt.

Consider refactoring the if not/continue structure in the while True
loop to avoid redundancy.