Refactor while True logic, rename/add functions.
Logic within the main while True loop is refactored to reduce
redundancy and increase readability.

Several functions are renamed to better reflect purpose.

presentFieldnames and confirmFieldnames functions are added to give user
ability to choose which fieldnames to include in the resulting PDF.
The logic is placed inside an additional while True loop to enable
reprompting after errors in user input are encountered and if user
chooses to reenter selections. Once confirmed, a message is displayed
in the terminal as with previous steps.