Add extractData function.
extractData takes in a file path and the user selected list of
fieldnames. This function uses CSV DictReader to parse the .csv file,
grab the data for selected fieldnames, and append this data in the form
dictionaries to a new list. That list of dicts is return and a
confirmation message is displayed in the terminal for the user.