Add createNewCSV function.
Creates path for a new CSV file to be temporarily stored
in user's Desktop folder, and opens this new file.

Using CSV DictWriter, it parses the passed in csv data and fieldnames,
writes the data to the new file, displays a confirmation message
in the terminal, and returns the new csv file path.