Add createPDF and removeBackupCSV functions.
createPDF creates a new PDF spreadsheet from the newly generated
CSV file, saved to the user Desktop. It then prints a confirmation
message to the terminal and returns the PDF file name.

removeBackupCSV deletes the temporary CSV file and displays
a confirmation message to the user.

Both functions are now called in main, and upon successful completion,
the PDF file will appear on the users Desktop.  A message is printed to
inform the user of the file name, location of the file, and reminder
to delete the file after use for security.