Add cleanData function to clean up uri values.
This function interates over a list, then over the dictionaries
within that list, and searches for a pattern match in the values.
It first checks to see if the value is a str before looking for the
pattern within that str.  If a pattern is found, it replaces the pattern
with match.group(1). Otherwise, the value remains the same. The item,
or key/value pair, is then appended to a new list.

This approach works regardless of fieldnames/keys, which are dynamically
determined in this program by the user at runtime.

Key benefits of adding this function are to save space on the PDF table,
increase readability of the PDF, and remove any sensitive data that may
have been previously stored in the uri data.