Add getFileName function
getFileName is a reusable function that takes in a file extension,
e.g. ".pdf" and asks user to input a name for their new file.

The function then checks that the user name contains the expected
extension, validates that the file name fits the valid pattern,
and returns the file name if valid. If not valid, the user is given
an error message and reprompted.