Add getTitle function and update createPDF
getTitle takes in a file_name to include in a confirmation message for
user.  This function asks user for a title to be included at the top of
each page in the new PDF document. The user is asked to confirm with
a 'Y' or 'N'. If 'N', the user will be prompted again for the title.
Otherwise, title is returned to be passed in to the createPDF function.